fix issues

+++ sensorslist.h
@@ -90,7 +90,7 @@ public:
     Q_INVOKABLE QString getPressure(int sensorId);
     Q_INVOKABLE int getXType(int sensorId);
 
-    Q_INVOKABLE int getProgressValue(int sensorId);
+    Q_INVOKABLE double getProgressValue(int sensorId);
     Q_INVOKABLE int getTempActiveValue(int sensorId);
     Q_INVOKABLE int getHeaterActiveValue(int sensorId);
     Q_INVOKABLE int getSensorActiveValue(int sensorId);
